Refactor movie fetching hooks to utilize a shared useMovieFetch hook for improved code reuse and maintainability
