+++ src/hooks/useNowPlayingMovies.ts
@@ -1,113 +1,8 @@
-import type React from "react";
-import { useCallback, useEffect, useState } from "react";
 import { api } from "../API";
-import type { MoviesState } from "./props";
-import { moviesState } from "./props";
-import { getCacheKey, getPersistedState, pruneSearchCache, setPersistedState } from "../helpers";
+import useMovieFetch, { type MovieFetchResult } from "./useMovieFetch";
 
-/**
- * Custom hook for fetching now playing movies.
- * It provides the current state of the movies, loading status, error status,
- * search term, and a function to load more movies.
- *
- * @returns {Object} - An object containing the current state of the movies, loading status, error status,
- * search term, and a function to load more movies.
- */
-function useNowPlayingMovies(): {
-  state: MoviesState;
-  loading: boolean;
-  error: boolean;
-  searchTerm: string;
-  setSearchTerm: React.Dispatch<React.SetStateAction<string>>;
-  setIsLoadingMore: React.Dispatch<React.SetStateAction<boolean>>;
-} {
-  const [error, setError] = useState(false);
-  const [isLoadingMore, setIsLoadingMore] = useState(false);
-  const [loading, setLoading] = useState(false);
-  const [searchTerm, setSearchTerm] = useState("");
-  const [state, setState] = useState(moviesState);
-
-  /**
-   * Fetches movies based on the provided page number and search term.
-   * If a search term is provided, it fetches movies matching the search term.
-   * Otherwise, it fetches now playing movies and sorts them by release date.
-   *
-   * @param {number} page - The page number of the movies to fetch.
-   * @returns {Promise<void>} - A promise that resolves when the movies are fetched.
-   */
-  const fetchMovies = useCallback(async (page: number, searchTerm = "") => {
-    try {
-      setError(false);
-      setLoading(true);
-
-      const movies = await api.fetchNowPlayingMovies(page, searchTerm);
-
-      movies.results.sort(
-        (a, b) =>
-          new Date(b.release_date).getTime() -
-          new Date(a.release_date).getTime(),
-      );
-
-      setState((prev) => ({
-        ...movies,
-        results:
-          page > 1 ? [...prev.results, ...movies.results] : movies.results,
-      }));
-    } catch (err) {
-      setError(true);
-      console.error("Failed to fetch now playing movies:", err);
-    } finally {
-      setLoading(false);
-    }
-  }, []);
-
-  useEffect(() => {
-    const load = async () => {
-      const cacheKey = getCacheKey("nowPlaying", searchTerm);
-
-      const cached = getPersistedState<typeof moviesState>(cacheKey);
-
-      if (cached) {
-        console.log("Grabbing from localStorage:", cacheKey);
-        setState(cached);
-        return;
-      }
-
-      console.log("Grabbing from API");
-      setState(moviesState);
-      await fetchMovies(1, searchTerm);
-    };
-
-    void load();
-  }, [searchTerm, fetchMovies]);
-
-  useEffect(() => {
-    if (!isLoadingMore) return;
-
-    const loadMore = async () => {
-      await fetchMovies(state.page + 1, searchTerm);
-      setIsLoadingMore(false);
-    };
-
-    void loadMore();
-  }, [fetchMovies, isLoadingMore, searchTerm, state.page]);
-
-  useEffect(() => {
-    if (state.results.length === 0) return;
-
-    setPersistedState(getCacheKey("nowPlaying", searchTerm), state);
-    
-    if (searchTerm) pruneSearchCache("nowPlaying");
-  }, [searchTerm, state]);
-
-  return {
-    state,
-    loading,
-    error,
-    searchTerm,
-    setSearchTerm,
-    setIsLoadingMore,
-  };
+function useNowPlayingMovies(): MovieFetchResult {
+  return useMovieFetch(api.fetchNowPlayingMovies, "nowPlaying", true);
 }
 
 export default useNowPlayingMovies;
